# cmd/api/main.go
package main

import (
	"log"

	"ayee-portal-backend/config"
	"ayee-portal-backend/internal/database"
	"ayee-portal-backend/internal/handlers"
	"ayee-portal-backend/internal/middleware"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "ayee-portal-backend/docs" // docs is generated by Swag CLI, you have to import it.
)

// @title           Ayee Portal API
// @version         1.0
// @description     This is the backend API for Ayee Portal.
// @termsOfService  http://swagger.io/terms/

// @contact.name    API Support
// @contact.url     http://www.swagger.io/support
// @contact.email   support@swagger.io

// @license.name    Apache 2.0
// @license.url     http://www.apache.org/licenses/LICENSE-2.0.html

// @host            localhost:8080
// @BasePath        /api

func main() {
	// 1. Load Config
	config.LoadConfig()

	// 2. Connect Database
	database.ConnectDB()

	// 3. Setup Router
	r := gin.Default()

	// CORS Middleware (Basic Example)
	r.Use(func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*") // Adjust for production
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE")
		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}
		c.Next()
	})

	// Swagger Route
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	api := r.Group("/api")
	{
		// Public Endpoints
		api.GET("/websites", handlers.GetWebsites)
		api.GET("/categories", handlers.GetCategories)
		api.GET("/settings", handlers.GetSystemSettings)
		// Assuming AuditLogs write might be public or protected? The user said "AuditLog... Create Website", usually audit logs are internal created by backend hooks or protected endpoints.
		// For now, let's make audit log creation protected or open if the frontend pushes it directly (less secure).
		// Given the local storage replacement context, the frontend probably pushes it.
		// But let's protect the write operations with AuthMiddleware
		
		protected := api.Group("/")
		protected.Use(middleware.AuthMiddleware())
		{
			protected.POST("/websites", handlers.CreateWebsite)
			protected.DELETE("/websites/:id", handlers.DeleteWebsite)
			
			protected.POST("/categories", handlers.CreateCategory)
			
			protected.GET("/audit-logs", handlers.GetAuditLogs)
			protected.POST("/audit-logs", handlers.CreateAuditLog) // If frontend reports actions

			protected.PUT("/settings", handlers.UpdateSystemSettings)

			protected.POST("/users/sync", handlers.SyncUser)
		}
	}

	port := config.AppConfig.ServerPort
	if port == "" {
		port = "8080"
	}

	log.Printf("Server running on port %s", port)
	if err := r.Run(":" + port); err != nil {
		log.Fatal("Failed to run server: ", err)
	}
}